move game related code to gametime file
