+++ lib/level.js
@@ -1,0 +1,11 @@
+function Level() {
+  this.columns = [];
+  this.rows = [];
+  this.paddleWidth = [];
+  this.ballRadius = [];
+  this.lives = [];
+  this.ballSpeed = [];
+  this.time = [];
+}
+
+module.exports = Level;
